+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/gfx/ExporterInfo.java
@@ -48,21 +48,7 @@ public class ExporterInfo extends Tag {
 
     public String swfName;
 
-    public List<Long> codeOffsets;
-
-    public static final int BITMAP_FORMAT_DEFAULT = 0;
-
-    public static final int BITMAP_FORMAT_TGA = 1;
-
-    public static final int BITMAP_FORMAT_DDS = 2;
-
-    //It looks like gfxexport produces BITMAP_FORMAT2_* values for format,
-    //but BITMAP_FORMAT_* works the same way
-    public static final int BITMAP_FORMAT2_JPEG = 10;
-
-    public static final int BITMAP_FORMAT2_TGA = 13;
-
-    public static final int BITMAP_FORMAT2_DDS = 14;
+    public List<Long> codeOffsets;  
 
     public static final int FLAG_CONTAINS_GLYPH_TEXTURES = 1;
 
@@ -139,26 +125,10 @@ public class ExporterInfo extends Tag {
         tagInfo.addInfo("general", "version", version);
         tagInfo.addInfo("general", "flags", "0x" + Long.toHexString(flags));
         String bitmapFormatStr = "0x" + Integer.toHexString(bitmapFormat);
-        switch (bitmapFormat) {
-            case BITMAP_FORMAT_DEFAULT:
-                bitmapFormatStr = "default (0)";
-                break;
-            case BITMAP_FORMAT_TGA:
-                bitmapFormatStr = "TGA (1)";
-                break;
-            case BITMAP_FORMAT_DDS:
-                bitmapFormatStr = "DDS (2)";
-                break;
-            case BITMAP_FORMAT2_JPEG:
-                bitmapFormatStr = "JPEG (10)";
-                break;
-            case BITMAP_FORMAT2_TGA:
-                bitmapFormatStr = "TGA (13)";
-                break;
-            case BITMAP_FORMAT2_DDS:
-                bitmapFormatStr = "DDS (14)";
-                break;
-        }
+        String fileFormatStr = FileFormatType.fileFormatToString(bitmapFormat);
+        if (fileFormatStr != null) {
+            bitmapFormatStr = fileFormatStr + " (" + bitmapFormat + ")";
+        } 
         tagInfo.addInfo("general", "bitmapFormat", bitmapFormatStr);
         tagInfo.addInfo("general", "prefix", prefix);
         tagInfo.addInfo("general", "swfName", swfName);