+++ CHANGELOG.md
@@ -25,6 +25,7 @@ All notable changes to this project will be documented in this file.
 - [#2243] Clean ABC action (remove unused items)
   available through context menu on ABC, ABCContainers, SWFs and in the ABC Explorer,
   `-abcclean` command on CLI
+- GFX - better fileformat detection
 
 ### Fixed
 - Debugger - getting children of top level variables